refactor: move ghost mode management to ghost-mode-manager
