+++ src/models/game.h
@@ -53,8 +53,4 @@ void freeGame();
 
 void goBackToMenu();
 
-void setGhostMode(GhostMode mode);
-GhostMode getGhostMode();
-void onGhostEatableTimerEnds();
-
 #endif // PACMAN_GAME_H
